Store files by sha256 instead of by path
Should be able to change a file name or path without having to redo
the work and leave a duplicate (with a different name) on the storage
endpoint.

Also remove some stream.end calls, because for some reason they cause
the files to be slightly larger or smaller?  Tested this a bunch and
appears to be fine without it, also does not appear to introduce a
leak or anything so?

This is most definitely a breaking change, so hopefully no-one but me
is using this :p